Move form back to theme
