+++ lib/project_types/theme/messages/messages.rb
@@ -47,6 +47,11 @@ module Theme
           },
           errors: "%s can't be blank",
           pull: {
+            ask_theme: "Select theme",
+            errors: {
+              invalid_password: "Invalid password",
+              connect_error: "Could not connect to given shop",
+            },
             private_app: <<~APP,
               To fetch your existing themes, Shopify App CLI needs to connect with your store. Visit {{underline:%s/admin/apps/private}} to create a new API key and password, or retrieve an existing password.
               If you create a new private app, ensure that it has Read and Write Theme access.
