--- lib/shopify-cli/forms.rb
@@ -1,7 +1,0 @@
-require 'shopify_cli'
-
-module ShopifyCli
-  module Forms
-    autoload :Pull, 'shopify-cli/forms/pull' # TODO: rename??
-  end
-end
